Extract Response.writeTo into separate class
And move response serialization test into ResponseWriterTests
